backend/requirements.txt
# Backend requirements for AWS Lambda functions
boto3==1.34.0
requests==2.31.0
pydantic==2.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0